--- apps/decodex/src/orchestrator/tests/recovery_runtime_reentry/exited_child/exited_child_reconciliation_detects_stalled_failed_runs_from_protocol_idle.rs
@@ -1,83 +1,0 @@
-use std::fs;
-
-use crate::{
-	agent::RUN_LEASE_IDLE_TIMEOUT,
-	orchestrator::{
-		self, RunLeaseDisposition,
-		tests::{self, FakeTracker},
-	},
-	state::{self, ProtocolActivityMarker, StateStore},
-};
-
-#[test]
-fn exited_child_reconciliation_detects_stalled_failed_runs_from_protocol_idle() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let issue = tests::sample_issue_with_sort_fields(
-		"issue-stalled-after-exit",
-		"PUB-205",
-		"In Progress",
-		&[],
-		Some(3),
-		"2026-03-13T04:16:17.133Z",
-	);
-	let tracker = FakeTracker::new(vec![issue.clone()]);
-	let run_id = "run-stalled-after-exit";
-	let worktree_path = config.worktree_root().join(&issue.identifier);
-
-	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
-
-	state_store
-		.record_run_attempt(run_id, &issue.id, 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			&issue.id,
-			"x/pubfi-pub-205",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-	state_store
-		.record_run_attempt(run_id, &issue.id, 1, "failed")
-		.expect("run should exit as failed before daemon inspects it");
-
-	state::write_run_protocol_activity_marker(
-		&worktree_path,
-		&ProtocolActivityMarker {
-			run_id,
-			attempt_number: 1,
-			thread_id: None,
-			turn_id: None,
-			event_count: 1,
-			last_event_type: "thread/status/changed",
-			child_agent_activity: None,
-			protocol_activity: None,
-		},
-	)
-	.expect("protocol marker should write");
-
-	let last_protocol_activity =
-		state::read_run_protocol_activity_marker(&worktree_path, run_id, 1)
-			.expect("protocol marker should read")
-			.expect("protocol activity should exist");
-	let actions = orchestrator::inspect_exited_daemon_child_reconciliation_at(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		&issue.id,
-		run_id,
-		last_protocol_activity + RUN_LEASE_IDLE_TIMEOUT.as_secs() as i64 + 1,
-	)
-	.expect("exited child inspection should succeed");
-
-	assert!(actions.iter().any(|action| {
-		action.issue.id == issue.id
-			&& matches!(
-				action.disposition,
-				RunLeaseDisposition::Stalled{ idle_for }
-					if idle_for >= RUN_LEASE_IDLE_TIMEOUT
-			)
-	}));
-}
