--- apps/decodex/src/orchestrator/tests/recovery_runtime_reentry/ownership_live_run/run_project_once_ignores_fresh_marker_for_exited_process.rs
@@ -1,37 +1,0 @@
-use crate::{
-	orchestrator::{
-		self,
-		tests::{self, FakeTracker, recovery_terminal_support},
-	},
-	state::{self, StateStore},
-	worktree::WorktreeManager,
-};
-
-#[test]
-fn run_project_once_ignores_fresh_marker_for_exited_process() {
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let issue = recovery_terminal_support::sample_active_issue("In Progress");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let worktree_manager =
-		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
-	let worktree = worktree_manager
-		.ensure_worktree(&issue.identifier, false)
-		.expect("recovered worktree should exist");
-	let exited_process_id = u32::MAX;
-
-	state::write_run_activity_marker_for_process(&worktree.path, "run-1", 1, exited_process_id)
-		.expect("activity marker should write");
-
-	let summary = orchestrator::run_project_once(&tracker, &config, &workflow, &state_store, true)
-		.expect("recovery should succeed")
-		.expect("dead process marker should not block retry planning");
-
-	assert_eq!(summary.issue_id, issue.id);
-	assert_eq!(summary.dispatch_mode, orchestrator::IssueDispatchMode::Retry);
-	assert!(
-		state_store.lease_for_issue(&issue.id).expect("lease lookup should succeed").is_none(),
-		"dead marker recovery should not reconstruct a live lease"
-	);
-}
